Stricter conversion checks.

+++ unit-tests/cl_mean.cpp
@@ -44,7 +44,7 @@ BOOST_AUTO_TEST_CASE(mean, * but::fixture(&setup))
   auto mat_f = mfmat::cl_mat_f{400, 500};
   for (std::size_t m = 0; m < mat_f.get_row_count(); ++m)
     for (std::size_t n = 0; n < mat_f.get_col_count(); ++n)
-      mat_f[{ m, n }] = m + n;
+      mat_f[{ m, n }] = static_cast<float>(m + n);
   auto res_f = mfmat::mean(mat_f);
   BOOST_CHECK_EQUAL(res_f.get_row_count(), 1);
   BOOST_CHECK_EQUAL(res_f.get_col_count(), 500);
@@ -58,7 +58,7 @@ BOOST_AUTO_TEST_CASE(mean, * but::fixture(&setup))
   auto mat_d = mfmat::cl_mat_d{700, 600};
   for (std::size_t m = 0; m < mat_d.get_row_count(); ++m)
     for (std::size_t n = 0; n < mat_d.get_col_count(); ++n)
-      mat_d[{ m, n }] = m + n;
+      mat_d[{ m, n }] = static_cast<double>(m + n);
   auto res_d = mfmat::mean(mat_d);
   BOOST_CHECK_EQUAL(res_d.get_row_count(), 1);
   BOOST_CHECK_EQUAL(res_d.get_col_count(), 600);
